requirements.txt
openai-whisper
streamlit
numpy
ffmpeg-python
transformers
yt-dlp
